process-i91/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-documentai
google-cloud-storage